feat: merge lighting systems
